Rename UnexpectedEmptyResponseError to UnexpectedHttpResponseError
