+++ app/errors/apiErrors/subclasses/UnexpectedHttpResponseError.test.js
@@ -1,33 +1,33 @@
 // @flow
 
-import { UnexpectedEmptyResponseError } from '../..';
+import { UnexpectedHttpResponseError } from '../..';
 
-describe(`UnexpectedEmptyResponseError`, (): void => {
+describe(`UnexpectedHttpResponseError`, (): void => {
 
   it(`can be thrown and caught`, (): void => {
     let error: typeof Error;
     try {
       // noinspection ExceptionCaughtLocallyJS
-      throw new UnexpectedEmptyResponseError('Test');
+      throw new UnexpectedHttpResponseError('Test');
     }
     catch (e) {
       error = e;
     }
 
-    expect(error).toBeInstanceOf(UnexpectedEmptyResponseError);
+    expect(error).toBeInstanceOf(UnexpectedHttpResponseError);
   });
 
   it(`sets a default message, if no message is passed`, (): void => {
     let error: typeof Error;
     try {
       // noinspection ExceptionCaughtLocallyJS
-      throw new UnexpectedEmptyResponseError();
+      throw new UnexpectedHttpResponseError();
     }
     catch (e) {
       error = e;
     }
 
-    expect(error).toBeInstanceOf(UnexpectedEmptyResponseError);
+    expect(error).toBeInstanceOf(UnexpectedHttpResponseError);
     expect((error: any).message).toBe(`Unexpected empty response data`);
   });
